+++ CHANGELOG.md
@@ -20,8 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   If you want to use node id printing, use `printHistory*(boolean showNodeId)` methods.
 - On `MoveNode` class (package-private, not exposed), removed not using `san` parameter.
 - On `MoveDataDTO` class, added `san` move string.
+- On every `nodeCounter` variables, changed `incrementAndGet` to `getAndIncrement`.
 
 ### Fixed
+- On `loadPGN` at `ChessGame`, fixed the `loadPGN` method didn't update node id counter.
 
 ### Performance
 - On `MoveGenerator`, replaced `if` cases on piece attack finding to `switch-case`.